refactor(i18n): move toLocalizationFileLoaderMap into  folder
